tllist: don’t implement tll_free() in terms of tll_remove()
tll_remove() is unnecessarily complex, as it updates the prev/next
pointers (and head/tail pointers, if necessary) of the surrounding
items.

Since we’re clearing the entire list, there’s no need for this.

So, simply iterate the list and “manually” call free(). Reset length,
head and tail once, at the end.